Wrap utils into object
